Updated cv and build script.

+++ scripts/build.sh
@@ -3,10 +3,8 @@
 format=${1:-tex_pdf}
 
 cd dist || exit
-if [ -f resume.pdf ]; then
-	json_resume convert --out="$format" --template=../templates/pdf.mustache ../rosshadden.json
-	mv resume.pdf rosshadden.pdf
-else
-	json_resume convert --out="$format" ../rosshadden.json
-fi
+
+json_resume convert --out="$format" --template=../templates/pdf.mustache ../rosshadden.json
+if [ -f resume.pdf ]; then mv resume.pdf rosshadden.pdf; fi
+
 cd .. || exit
